Fixed the Reader Mode Problem and Extracted Extension Reader Mode Properties
