+++ Client/Frontend/Browser/BrowserViewController/BrowserViewController+ReaderMode.swift
@@ -1,0 +1,207 @@
+// Copyright 2020 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import Shared
+import WebKit
+import Storage
+
+// MARK: - ReaderModeDelegate
+
+extension BrowserViewController: ReaderModeDelegate {
+    func readerMode(_ readerMode: ReaderMode, didChangeReaderModeState state: ReaderModeState, forTab tab: Tab) {
+        // If this reader mode availability state change is for the tab that we currently show, then update
+        // the button. Otherwise do nothing and the button will be updated when the tab is made active.
+        if tabManager.selectedTab === tab {
+            topToolbar.updateReaderModeState(state)
+        }
+    }
+
+    func readerMode(_ readerMode: ReaderMode, didDisplayReaderizedContentForTab tab: Tab) {
+        showReaderModeBar(animated: true)
+        tab.showContent(true)
+    }
+
+    func readerMode(_ readerMode: ReaderMode, didParseReadabilityResult readabilityResult: ReadabilityResult, forTab tab: Tab) { }
+}
+
+// MARK: - ReaderModeStyleViewControllerDelegate
+
+extension BrowserViewController: ReaderModeStyleViewControllerDelegate {
+    func readerModeStyleViewController(_ readerModeStyleViewController: ReaderModeStyleViewController, didConfigureStyle style: ReaderModeStyle) {
+        // Persist the new style to the profile
+        let encodedStyle: [String: Any] = style.encodeAsDictionary()
+        profile.prefs.setObject(encodedStyle, forKey: ReaderModeProfileKeyStyle)
+        // Change the reader mode style on all tabs that have reader mode active
+        for tabIndex in 0..<tabManager.count {
+            if let tab = tabManager[tabIndex] {
+                if let readerMode = tab.getContentScript(name: "ReaderMode") as? ReaderMode {
+                    if readerMode.state == ReaderModeState.active {
+                        readerMode.style = style
+                    }
+                }
+            }
+        }
+    }
+}
+
+// MARK: - ReaderModeBarViewDelegate
+
+extension BrowserViewController: ReaderModeBarViewDelegate {
+    
+    func readerModeBar(_ readerModeBar: ReaderModeBarView, didSelectButton buttonType: ReaderModeBarButtonType) {
+        switch buttonType {
+        case .settings:
+            if let readerMode = tabManager.selectedTab?.getContentScript(name: "ReaderMode") as? ReaderMode, readerMode.state == ReaderModeState.active {
+                var readerModeStyle = DefaultReaderModeStyle
+                if let dict = profile.prefs.dictionaryForKey(ReaderModeProfileKeyStyle) {
+                    if let style = ReaderModeStyle(dict: dict as [String: AnyObject]) {
+                        readerModeStyle = style
+                    }
+                }
+
+                let readerModeStyleViewController = ReaderModeStyleViewController()
+                readerModeStyleViewController.delegate = self
+                readerModeStyleViewController.readerModeStyle = readerModeStyle
+                readerModeStyleViewController.modalPresentationStyle = .popover
+
+                let setupPopover = { [unowned self] in
+                    if let popoverPresentationController = readerModeStyleViewController.popoverPresentationController {
+                        popoverPresentationController.backgroundColor = UIColor.Photon.white100
+                        popoverPresentationController.delegate = self
+                        popoverPresentationController.sourceView = readerModeBar
+                        popoverPresentationController.sourceRect = CGRect(x: readerModeBar.frame.width/2, y: UIConstants.toolbarHeight, width: 1, height: 1)
+                        popoverPresentationController.permittedArrowDirections = .up
+                    }
+                }
+
+                setupPopover()
+
+                if readerModeStyleViewController.popoverPresentationController != nil {
+                    displayedPopoverController = readerModeStyleViewController
+                    updateDisplayedPopoverProperties = setupPopover
+                }
+
+                present(readerModeStyleViewController, animated: true, completion: nil)
+            }
+        }
+    }
+}
+
+// MARK: - ReaderModeBarUpdate
+
+extension BrowserViewController {
+    
+    func updateReaderModeBar() {
+        if let readerModeBar = readerModeBar {
+            let theme = Theme.of(tabManager.selectedTab)
+            readerModeBar.applyTheme(theme)
+        }
+    }
+
+    func showReaderModeBar(animated: Bool) {
+        if self.readerModeBar == nil {
+            let readerModeBar = ReaderModeBarView(frame: CGRect.zero)
+            readerModeBar.delegate = self
+            view.insertSubview(readerModeBar, belowSubview: header)
+            self.readerModeBar = readerModeBar
+            
+            readerModeBar.snp.makeConstraints { make in
+                make.top.equalTo(self.header.snp.bottom)
+                make.height.equalTo(UIConstants.toolbarHeight)
+                make.leading.trailing.equalTo(self.view)
+            }
+        }
+
+        updateReaderModeBar()
+
+        updateViewConstraints()
+    }
+
+    func hideReaderModeBar(animated: Bool) {
+        if let readerModeBar = self.readerModeBar {
+            readerModeBar.removeFromSuperview()
+            self.readerModeBar = nil
+            self.updateViewConstraints()
+        }
+    }
+
+    /// There are two ways we can enable reader mode. In the simplest case we open a URL to our internal reader mode
+    /// and be done with it. In the more complicated case, reader mode was already open for this page and we simply
+    /// navigated away from it. So we look to the left and right in the BackForwardList to see if a readerized version
+    /// of the current page is there. And if so, we go there.
+
+    func enableReaderMode() {
+        guard let tab = tabManager.selectedTab, let webView = tab.webView else { return }
+
+        let backList = webView.backForwardList.backList
+        let forwardList = webView.backForwardList.forwardList
+
+        guard let currentURL = webView.backForwardList.currentItem?.url, let readerModeURL = currentURL.encodeReaderModeURL(WebServer.sharedInstance.baseReaderModeURL()) else { return }
+
+        if backList.count > 1 && backList.last?.url == readerModeURL {
+            webView.go(to: backList.last!)
+        } else if forwardList.count > 0 && forwardList.first?.url == readerModeURL {
+            webView.go(to: forwardList.first!)
+        } else {
+            // Store the readability result in the cache and load it. This will later move to the ReadabilityHelper.
+            webView.evaluateJavaScript("\(ReaderModeNamespace).readerize()", completionHandler: { (object, error) -> Void in
+                if let readabilityResult = ReadabilityResult(object: object as AnyObject?) {
+                    try? self.readerModeCache.put(currentURL, readabilityResult)
+                    if let nav = webView.load(PrivilegedRequest(url: readerModeURL) as URLRequest) {
+                        self.ignoreNavigationInTab(tab, navigation: nav)
+                    }
+                }
+            })
+        }
+    }
+
+    /// Disabling reader mode can mean two things. In the simplest case we were opened from the reading list, which
+    /// means that there is nothing in the BackForwardList except the internal url for the reader mode page. In that
+    /// case we simply open a new page with the original url. In the more complicated page, the non-readerized version
+    /// of the page is either to the left or right in the BackForwardList. If that is the case, we navigate there.
+
+    func disableReaderMode() {
+        if let tab = tabManager.selectedTab,
+            let webView = tab.webView {
+            let backList = webView.backForwardList.backList
+            let forwardList = webView.backForwardList.forwardList
+
+            if let currentURL = webView.backForwardList.currentItem?.url {
+                if let originalURL = currentURL.decodeReaderModeURL {
+                    if backList.count > 1 && backList.last?.url == originalURL {
+                        webView.go(to: backList.last!)
+                    } else if forwardList.count > 0 && forwardList.first?.url == originalURL {
+                        webView.go(to: forwardList.first!)
+                    } else {
+                        if let nav = webView.load(URLRequest(url: originalURL)) {
+                            ignoreNavigationInTab(tab, navigation: nav)
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    @objc func dynamicFontChanged(_ notification: Notification) {
+        guard notification.name == .dynamicFontChanged else { return }
+
+        var readerModeStyle = DefaultReaderModeStyle
+        if let dict = profile.prefs.dictionaryForKey(ReaderModeProfileKeyStyle) {
+            if let style = ReaderModeStyle(dict: dict as [String: AnyObject]) {
+                readerModeStyle = style
+            }
+        }
+        readerModeStyle.fontSize = ReaderModeFontSize.defaultSize
+        self.readerModeStyleViewController(ReaderModeStyleViewController(), didConfigureStyle: readerModeStyle)
+    }
+    
+    func ignoreNavigationInTab(_ tab: Tab, navigation: WKNavigation) {
+        self.ignoredNavigation.insert(navigation)
+    }
+    
+    func recordNavigationInTab(_ tab: Tab, navigation: WKNavigation, visitType: VisitType) {
+        self.typedNavigation[navigation] = visitType
+    }
+}
